+++ R/screening_classes.R
@@ -1,3 +1,9 @@
+#' @keywords internal
+update_screen_default <- function(object, x, y, family, ...) {
+  object
+}
+
+
 #' Constructor Function for Building \code{'screencoef'} Objects
 #'
 #' Creates an object class \code{'screencoef'} using arguments passed by user.
@@ -6,6 +12,14 @@
 #'    function should have arguments  and   \code{y} (vector of responses -- standardized
 #'    for Gaussian family), \code{x} (the matrix of standardized predictors) and a
 #'    \code{'screencoef'} object.
+#' @param update_fun optional function for updating the \code{'screencoef'} object with
+#' information from the data passed to `spar()`. This
+#' function should have arguments \code{object}, which is a \code{'randomprojection'}
+#' object, `x`, y` (the predictor matrix and response vector supplied by `spar()`.
+#' Both have already been standardized by `spar()`),
+#' `family` and `...`, whereas all other potentially relevant arguments of `spar()`
+#' are passed internally to this function through `...`.
+#' If `update_fun` is not provided, the object remains unchanged.
 #' @return a function which in turn creates an object of class \code{'screencoef'}
 #' @description
 #' The created function will return a object of class \code{'screencoef'} which
@@ -15,12 +29,12 @@
 #' \code{"fixed"} (indicating that the top \code{nscreen} variables should be employed).
 #' @examples
 #' generate_scr_sirs <- function(y, x, object) {
-#'   res_screen <- do.call(function(...)
-#'     VariableScreening::screenIID(x, y, ...),
-#'     object$control)
-#'   coefs <- res_screen$measurement
-#'   coefs
-#' }
+#'   ctrl <- object$control[names(object$control) %in%
+#'                           names(formals(VariableScreening::screenIID))]
+#'  res_screen <- do.call(function(...)
+#'    VariableScreening::screenIID(x, y, ...), ctrl)
+#'  res_screen$measurement
+#'}
 #' screen_sirs <- constructor_screencoef("screen_sirs",
 #'   generate_fun = generate_scr_sirs)
 #' example_data <- simulate_spareg_data(n = 100, p = 400, ntest = 100)
@@ -29,7 +43,8 @@
 #'   rp = rp_sparse())
 #' spar_example
 #' @export
-constructor_screencoef <- function(generate_fun, name = NULL) {
+constructor_screencoef <- function(name = NULL, generate_fun,
+                                   update_fun = update_screen_default) {
   ## Checks
   args_generate_fun <- formals(generate_fun)
   stopifnot("Function generate_fun should contain three arguments: x, y and an object
@@ -39,10 +54,12 @@ constructor_screencoef <- function(generate_fun, name = NULL) {
               "y" %in% names(args_generate_fun))
   stopifnot("Function generate_fun should contain argument 'x', the matrix of predictors." =
               "x" %in% names(args_generate_fun))
+  stopifnot("Function update_fun should have as argument object, x, y, family, ..." = names(formals(update_fun)) %in% c("object","x", "y", "family", "..."))
   ## Function to return
   function(..., control = list()) {
     out <- list(name = name,
                 generate_fun = generate_fun,
+                update_fun = update_fun,
                 control = control)
     attr <- list2(...)
     attributes(out) <- c(attributes(out), attr)
@@ -60,6 +77,11 @@ constructor_screencoef <- function(generate_fun, name = NULL) {
   }
 }
 
+#' @keywords internal
+update_screen_marglik <- function(object, x, y, family, ...) {
+  if (is.null(object$control$family)) object$control$family <- family
+  object
+}
 #'
 #' Generate screening coefficient based  on marginal likelihood in univariate GLMs
 #' @param y vector of responses
@@ -68,17 +90,16 @@ constructor_screencoef <- function(generate_fun, name = NULL) {
 #' @return vector of screening coefficients of length p
 #' @keywords internal
 generate_scrcoef_marglik <- function(y, x, object) {
-  control <- object$control
-  if (is.null(control$family)) {
-    control$family <- eval(parse(text=attr(object, "family_string")))
-  }
+  ctrl <- object$control[names(object$control) %in%
+                         names(formals(glm))]
   coefs <- apply(x, 2, function(xj){
     glm_res <- do.call(function(...) glm(y ~ xj,  ...),
-                       control)
+                       ctrl)
     glm_res$coefficients[2]
   })
   coefs
 }
+
 #' Screening Coefficient Based on Marginal GLMs
 #'
 #' @param ... includes arguments which can be passed as attributes to the
@@ -135,7 +156,8 @@ generate_scrcoef_marglik <- function(y, x, object) {
 #'
 screen_marglik <- constructor_screencoef(
   name = "screen_marglik",
-  generate_fun = generate_scrcoef_marglik)
+  generate_fun = generate_scrcoef_marglik,
+  update_fun = update_screen_marglik)
 
 
 #'
@@ -153,6 +175,9 @@ generate_scrcoef_cor <- function(y, x, object) {
   })
   coefs
 }
+
+
+
 #' Screening Coefficient Based on Correlation
 #'
 #' Creates an object class \code{'screencoef'} using arguments passed by user.
@@ -209,6 +234,65 @@ screen_cor <- constructor_screencoef(
   name = "screen_cor",
   generate_fun = generate_scrcoef_cor)
 
+
+#'
+#' Update screening glmnet object
+#' @param object  \code{'screencoef'} object
+#' @param y vector of responses
+#' @param x matrix of predictors
+#' @param family family object to be passed from \code{spar()}
+#' @return  vector of screening coefficients of length p
+#' @keywords internal
+update_screen_glmnet <- function(object, x, y, family, ...) {
+   # Family compatibility with glmnet
+  if (is.null(object$control$family)) {
+    object$control$family <-  eval(parse(text=attr(object, "family_string")))
+  }
+#   if (!is.null(object$control$family)) {
+#     fam <- object$control$family
+#   } else {
+#     fam <- family
+#   }
+#   if (fam$family == "gaussian" & fam$link=="identity") {
+#     fit_family <- "gaussian"
+#   } else {
+#     if (fam$family=="binomial" & fam$link=="logit") {
+#       fit_family <- "binomial"
+#     } else if (fam$family=="poisson" & fam$link=="log") {
+#       fit_family <- "poisson"
+#     } else {
+#       fit_family <- fam
+#     }
+#   }
+#   object$control$family <- family# fit_family
+  # Set alpha by default to 0
+  if (is.null(object$control$alpha)) object$control$alpha <- 0
+  # Set default for lambda.min.ratio
+  fam <- object$control$family
+  if (is.null(object$control$lambda.min.ratio)) {
+    n <- NROW(x)
+    tmp_sc <- apply(x, 2, function(col) sqrt(var(col)*(n-1)/n))
+    # TODO: allow for weights in the future
+    x2 <- scale(x, center = colMeans(x), scale = tmp_sc)
+    mu0 <- glm(y ~ 1, family = fam)$fitted.values
+    r <- y - mu0
+    eta <- fam$linkfun(mu0)
+    v <- fam$variance(mu0)
+    me <- fam$mu.eta(eta)
+    rv <- 1/n * r / v * me
+    lam_max <- 1000 * max(abs(crossprod(rv, x2[,tmp_sc > 0])))
+    object$control$lambda.min.ratio <- min(0.01, 1e-4 / lam_max)
+  }
+
+  # Set cutoff ratio for deviance
+   if (is.null(object$control$dev.ratio_cutoff)) {
+     object$control$dev.ratio_cutoff <- ifelse(fam$family == "gaussian", 0.999, 0.8)
+   }
+#
+   object
+#
+}
+
 #'
 #' Screening coefficient based  on glmnet coefficients
 #' @param y vector of responses
@@ -217,42 +301,14 @@ screen_cor <- constructor_screencoef(
 #' @return vector of screening coefficients of length p
 #' @keywords internal
 generate_scrcoef_glmnet <- function(y, x, object) {
-  n <- NROW(x)
-  p <- NCOL(x)
   control_glmnet <-
-    object$control[names(object$control)  %in% names(formals(glmnet))]
+    object$control[names(object$control) %in% names(formals(glmnet))]
 
-  if (is.null(control_glmnet$family)) {
-    control_glmnet$family <-  eval(parse(text=attr(object, "family_string")))
-  }
-  family <- control_glmnet$family
-
-  # Set alpha to zero unless otherwise specified
-  if (is.null(control_glmnet$alpha)) control_glmnet$alpha <- 0
-  # Set lambda.min.ration to close to zero unless otherwise specified
-  if (is.null(control_glmnet$lambda.min.ratio)) {
-    tmp_sc <- apply(x, 2, function(col) sqrt(var(col)*(n-1)/n))
-    # TODO: allow for weights in the future
-    x2 <- scale(x, center = colMeans(x), scale = tmp_sc)
-    mu0 <- glm(y ~ 1, family = family)$fitted.values
-    r <- y - mu0
-    eta <- family$linkfun(mu0)
-    v <- family$variance(mu0)
-    me <- family$mu.eta(eta)
-    rv <- 1/n * r / v * me
-    lam_max <- 1000 * max(abs(crossprod(rv, x2[,tmp_sc > 0])))
-    control_glmnet$lambda.min.ratio <- min(0.01, 1e-4 / lam_max)
-  }
-  # Obtain Ridge coefs GLMNET
+  # Obtain penalized coefs GLMNET
   glmnet_res <- do.call(function(...) glmnet(x = x, y = y, ...),
                         control_glmnet)
 
-  if (family$family == "gaussian") {
-    dev.ratio_cutoff <- 0.999
-  } else {
-    dev.ratio_cutoff <- 0.8
-  }
-  lam <- min(glmnet_res$lambda[glmnet_res$dev.ratio <= dev.ratio_cutoff])
+  lam <- min(glmnet_res$lambda[glmnet_res$dev.ratio <= object$control$dev.ratio_cutoff])
   scr_coef <- coef(glmnet_res, s = lam)[-1]
   scr_coef
 }
@@ -312,7 +368,8 @@ generate_scrcoef_glmnet <- function(y, x, object) {
 #'
 screen_glmnet <- constructor_screencoef(
   name = "screen_glmnet",
-  generate_fun = generate_scrcoef_glmnet)
+  generate_fun = generate_scrcoef_glmnet,
+  update_fun = update_screen_glmnet)
 
 
 #' Print Method for \code{'screencoef'} Object