+++ R/utilities.R
@@ -132,3 +132,4 @@ compute_val_summary <- function(val_res) {
   return(val_sum)
 }
 
+